refactor: 💡 optimize block handle code
